refactor: features/calendar 리팩토링
