+++ src/entities/calendar/config/constants.ts
@@ -1,7 +1,5 @@
 import { generateExampleGuideSchedules } from '../lib/helpers'
 
-export const DAY_OF_THE_WEEK = ['일', '월', '화', '수', '목', '금', '토']
-
 export const COLORS = {
   LAVENDER: '#9a9cff',
   RED: '#dc2227',
