+++ src/features/calendar/lib/helpers.ts
@@ -11,7 +11,39 @@ import {
   startOfDay,
   startOfMonth,
 } from 'date-fns'
-import { DAY_OF_THE_WEEK } from '../../../entities/calendar/config/constants'
+
+/**
+ * @param index
+ * @param allIndices // 일정의 index 배열 ex) [0, 1, 2]
+ * @returns
+ */
+export const getAdjustedIndex = (index: number, allIndices: number[]): number => {
+  // index가 3보다 작으면 그대로 반환
+  if (index < 3) return index
+
+  // allIndices에서 3보다 작은 index를 필터링
+  const usedIndices = allIndices.filter((idx) => idx < 3)
+
+  // allIndices에서 3보다 큰 index를 필터링하고 정렬
+  const highIndices = allIndices.filter((idx) => idx >= 3).sort((a, b) => a - b)
+
+  // highIndices에서 현재 index의 위치 찾기
+  const highIndexPosition = highIndices.indexOf(index)
+
+  // availableIndices (0,1,2)에서 사용되지 않은 index 찾기
+  const availableIndices = [0, 1, 2].filter((idx) => !usedIndices.includes(idx))
+
+  // availableIndices가 있고 highIndexPosition이 -1이 아니면
+  if (availableIndices.length > 0 && highIndexPosition !== -1) {
+    // availableIndices의 순서대로 할당
+    if (highIndexPosition < availableIndices.length) {
+      return availableIndices[highIndexPosition]
+    }
+  }
+
+  // 할당이 불가능하면 원본 index 반환
+  return index
+}
 
 interface GenerateDateProps {
   month: number
@@ -54,23 +86,6 @@ export const generateDate = ({ year, month }: GenerateDateProps) => {
   return arrayOfDate
 }
 
-export const getFormattedDate = (date: Date) => {
-  return date.getDate()
-}
-
-export const getFormattedDay = (date: Date) => {
-  const day = date.getDay()
-  return DAY_OF_THE_WEEK[day]
-}
-
-export const getFormattedTime = (date: string) => {
-  const time = date.split('T')[1].split(':')
-
-  const hour = time[0]
-  const minute = time[1]
-  return `${hour}:${minute}`
-}
-
 export const getOneDaySchedule = (schedules: ScheduleResponse[], date: string | Date) => {
   // console.log('성능 최적화가 되나?')
   if (schedules.length === 0) return []
@@ -81,11 +96,6 @@ export const getOneDaySchedule = (schedules: ScheduleResponse[], date: string | 
       end: endOfDay(plan?.endDateTime),
     }),
   )
-}
-
-// 오늘이 일정 시작일인지 체크
-export const isStartDate = (startDate: string, today: Date) => {
-  return isSameDay(startDate, today)
 }
 
 // 1. 일정 시작일이 빠른 순서로 정렬
@@ -112,4 +122,9 @@ export const trancateString = (str: string, maxLength: number) => {
 // 하루짜리 일정인지 확인
 export const isShortPlan = (startDate: string, endDate: string) => {
   return isSameDay(startDate, endDate)
+}
+
+// 오늘이 일정 시작일인지 체크
+export const isStartDate = (startDate: string, today: Date) => {
+  return isSameDay(startDate, today)
 }
