+++ src/shared/config/constants.ts
@@ -2,3 +2,5 @@ export const ACCESS_TOKEN_KEY = 'access_token'
 export const REFRESH_TOKEN_KEY = 'refresh_token'
 
 export const LOGIN_SUCCESS = 'login_success'
+
+export const DAY_OF_THE_WEEK = ['일', '월', '화', '수', '목', '금', '토']
