+++ src/shared/lib/index.ts
@@ -4,3 +4,4 @@ export { useRequireAuth } from './hooks/useRequireAuth'
 export { cookiesUtil } from './utils/cookies'
 export { localStorageUtil } from './utils/localStorage'
 export { sessionStorageUtil } from './utils/sessionStorage'
+export * from './utils/date'
